Switch FFI library to `ffi-napi`
This switches the underlying library used for FFI calls from
`ffi` to `ffi-napi`, a port of `ffi` to Node’s new N-API
addon interface.

This is happening as a larger effort to increase N-API
usage in the ecosystem as a whole.

The benefits of this change for this library are:

- As it stands, `ffi` flat out doesn’t even build on Node 9
  and above – `ffi-napi` does. A patch for that has been merged
  but new releases of `ffi` have not been cut in over a year,
  and generally maintenance has seen a noticeable decrease over time.
- As with any library using N-API for their native bindings,
  big advantages are that these libraries need not be recompiled
  when a new version of Node is used, and that it is completely
  engine-independent (i.e. this code is going to work with ChakraCore
  as well, without rewriting or even just recompiling).

The only (possible) downside would be that N-API is still a
relatively new technology and that `ffi-napi` itself is still quite
young as well, so there is a slight chance of running into issues
with either technology. Just as a data point, this commit is
passing all tests on all supported versions of Node.js.

Generally, me and the Node.js N-API team would be happy to see adoption
increase and learn about any difficulties that people are running
into during that process and would really like to help address those
issues.

I hope you consider this change!